Add `FlattenAs<Pack, Ts...>` helper
This will take any and all instances of `Pack<...>`, regardless of
nesting, and replace them with the contents of the pack.  All results
will be wrapped in one overarching `Pack<...>` instance.

The point of this is to make it easy to keep the future `OpSequence` in
a canonical form.  We could easily have an `OpSequence` as one of the
elements of an `OpSequence`.  This helper will let us avoid having weird
combinations persist.  This helps pave the way for #349.