Handle lowest signed integrals in `get_value` (#447)
Currently, we can't represent `-mag<128>()` in `int8_t`, even though the
value `-128` _is_ representable.  The problem is that we're using the
max of the type to decide what we can represent, and then just negating
it.  This special case logic enables us to handle all integer values
correctly.

Helps pave the way for #349.