Add upper-bound overflow checker for `StaticCast`
This completes our overflow checking capabilities for `StaticCast`, in
the obvious way.

Note that we have a different set of subtleties to wrestle with than for
the min.  For the min, we had to deal with unsigned types having a lower
bound of 0, although the signed types' lower bounds were nice powers of
two.  For the max, in this PR, we have to watch out for the upper bounds
being _slightly lower_ than a nice power of two.  This has surprising
consequences, such as the fact that you can cast an int type's max to a
floating point type, but you _can't cast it back_.  We deal with these
as best we can, bearing in mind that many common tools for dealing with
floating point types are not available in `constexpr` for C++14.

Helps #349.