Add abstract operations (#450)
We add special types for the kinds of operations that can occur in unit
conversions.  Later on, we'll create traits that judge the risks of
these operations for overflow and truncation, and can even determine
which individual values will be lossy in these ways.

For now, the operations are:

- `StaticCast<T, U>` represents a static cast from type `T` to `U`.
- `MultiplyTypeBy<T, M>` represents multiplying a value of `T` by the
  magnitude `M`.
- `DivideTypeByInteger<T, M>` represents dividing a value of `T` by the
  magnitude `M`, which must be an integer.
- `OpSequence<Ops...>` chains any sequence of operations, and is
  automatically flattened to a single layer.

We also provide the `OpInput<Op>` and `OpOutput<Op>` traits.  For
`OpSequence`, we also explicitly check that the output type of one
operation equals the input type of the next.  This has helped prevent a
few errors already.

Helps #349.