Add lower-bound overflow checker for `StaticCast` (#452)
The new `:overflow_boundary` target will check any operation (from
`:abstract_operations`) and find the minimum and maximum acceptable
value.  Here, we are starting with `StaticCast` only, and `MinGood`
only, to get us up and running.  (The final implementation for all
operations is currently almost 900 lines of code and almost 1600 lines
of tests!)

Strategically, we're thinking about non-arithmetic reps too (#52), even
though we're not ready to actually support them formally.  The strategy
will be to ask any supplier of a non-arithmetic rep to carefully
specialize `std::numeric_limits`.  Then, we'll rely on the answers they
give to judge overflow or truncation risk.  If they don't give answers,
we'll just assume conversions are risky.  `std::numeric_limits` is
probably pretty flawed in general, but I think this will be far more
workable than asking users of a non-arithmetic rep to specialize Au
specific traits.

One other important distinction is that these traits don't return their
values in `OpInput<T>`, but in `RealPart<OpInput<T>>`.  If we think
about, say, a full complex number, then notions of "max" and "min" don't
even make sense, because complex numbers are not ordered!  In fact, we
expect to replace `RealPart<T>` with a more general notion of
`ScalarPart<T>` in order to support #70, because the same considerations
apply to vectors and matrices.  But we will save that for when we are
ready to tackle #70 in earnest.

Helps #349.