Implement `MinGood<MultiplyTypeBy<T, M>>` (#455)
If `T` is unsigned, this should always be 0 (because 0 is the minimum
_possible_, and 0 is _always_ "good").  Otherwise, we check whether `M`
is "the sort of thing" that can be represented in `T`.  If not, well, we
just return 0 as a placeholder.  If it _is_ compatible, we _divide_ the
limits by the magnitude, and take whichever is lower --- unless the
result would be lower than the minimum value of the type, in which case
we clamp it.

Helps #349.