Implement limits for `OpSequence`
Now it all comes together.  We carefully designed the overflow checkers
for each individual operation so that they took a `Limits` parameter,
which might be different from the full limits for the type.  This PR
adds a `LimitsFor<Op, Limits>` helper, which takes an "output target"
limits type `Limits`, and an operation `Op`, and produces the _input
limits_ that will stay within that target.  And this object can be used
as the "limits" parameter _for the previous operation in the chain_!
And we're done.

Helps #349.