Implement `MaxGood<MultiplyTypeBy<T, M>>` (#456)
The main hurdle here is that for signed integral types, the minimum
value is more negative than the max value is positive.  To deal with
this, we add a `clamped_negate(T)` function, which negates the input,
but saturates at the limits of the type.  We use it consistently in a
few other places where it makes sense.

We also add an `equal` function to the `MagHelper`, so that we can
seamlessly check whether a value in a given type happens to equal a
`Magnitude`.  This again helps us deal with this hurdle that we have
with signed integral types.

With these helpers in hand, we can implement `MaxGood` using a very
similar strategy as for `MinGood`.  The only substantive difference is
that `MinGood` automatically returns `0` for all unsigned types period,
but `MaxGood` only does this for an unsigned type _with a negative scale
factor_.

Helps #349.