Add foundations for `:truncation_risk` target
This adds one (templated) type for each of various kinds of truncation
risk that we might have.  Each type has a `would_value_truncate(T)`
member, which assesses _any individual value_ of type `T` to see if it
would truncate.  Future PRs will associate each operation with one of
these.

This set isn't perfect.  In particular, I'm pretty suspicious as to
whether `ValueTimesRatioIsNotInteger` is the right solution.  But in my
local client, this suite seems to be able to handle all the cases I
checked.  I think **the most important thing is the structure**.  If we
land a suite that passes all known test cases, then we can clean up the
particular details later on, and trust in our expanding set of test
cases to be sure we end up in a better place.

Helps #349.  Technically, that's only about overflow, but we can't
actually _use_ the new overflow implementations without moving to this
operation based approach, which means we need to be able to handle
truncation too.  Besides, this will ultimately lead to more efficient
truncation checkers as well.