Add fuzz-style explorer for conversion lossiness
This is a simple tool to randomly generate values in a wide variety of
types, and try applying a bunch of unit conversions.  We make sure that
the round trip value changes if and only if we expect the conversion to
be lossy.

Of course, there are a variety of edge case ways for this to be
mismatched that we deem to be OK.  For example, floating point values
can be "subnormal", and we want to consider those out of scope.
Additionally, we assume that a floating point conversion can differ from
the best value by up to 1 ULP, so a round trip can be off by up to 2
ULP.  There are a few other carveouts of this nature.

If we find a mismatch, we terminate and immediately print out details
about the specific test case.  The expectation is that we will either
add a test case to the library and fix the implementation, or else add a
carveout to this utility.  The goal is to get to a state where we can
run continuously and not find any unaccepted anomalies.

Overall, this tool should greatly increase our confidence that our
runtime overflow and truncation checkers are working correctly.

Helps #349.