Add `:conversion_strategy` target (#460)
This new target is the One True Home for choosing each unit conversion
implementation.  It depends only on the initial rep, final rep, and
conversion factor.  The output is a tailor-made and trimmed sequence of
operations for that conversion.  It will automatically take care of:

- Type promotion
- Using division for pure-inverse-integer factors
- Using multiply-then-divide for applying a rational to an integral type

When all is said and done for #349, each conversion implementation will
be able to simply ask for the right conversion for the reps and factor,
and just apply it blindly.  And when we want to reason about overflow
and/or truncation risk for those conversions, we can do so by checking
this sequence of operations!