Add runtime overflow checker
The interface is `would_value_overflow<Op>(x)`, where `x` is a value of
type `OpInput<Op>`.  This interface is already good enough to replace
both the implicit-rep and explicit-rep versions of the runtime overflow
checkers!  Except that we can't just yet, because we're not _actually
using an operation_ to implement our conversions.

Along the way, we add (and test) some other helpful utilities, some of
which are useful in other places.  For example, we'll make use of
`CanOverflowAbove` in our implicit conversion policy.

`MinValueChecker` and `MaxValueChecker` are designed to produce
functions that unconditionally return `false` for their `is_too_XX()`
members.  This should make it very easy for the compiler to produce very
efficient assembly for `would_value_overflow()`.

Helps #349: at this point, the `:overflow_boundary` target should be
complete and ready for production!