Assess truncation risk for every simple operation (#462)
This means adding a new trait, `TruncationRiskFor<Op>`, which maps any
operation onto one of the truncation risk types that we added in the
previous PR.

These "simple" operations are reasonably straightforward, so I'm adding
all of them in a single PR.  Handling operation _sequences_ will be more
involved, and I expect it to take two more PRs.

Helps #349.