Add `UpdateRisk<Op, Risk>` helper (#468)
Our goal is to assess truncation risk for `OpSequence<Ops...>`
instances.  Just like with overflow risk, we'll start from the last
operation, and work our way backwards.  But the substance is quite
different.

First: we are making the explicit assumption that we can always reduce
the risk for a compound operation to a single risk category.  This may
or may not always be strictly true.  However, it does seem like it's at
least good enough to get a working implementation that we can start
with.  And there's a good chance that it _is_ true if we restrict
ourselves to _the kinds of operation sequences that we will actually
use_ in unit conversions.

(For posterity: what would we do if we learned that this assumption is
wrong?  We would return a _list_ of risks, and assess all of them.  This
list should almost always turn out to have only one element anyway.)

Now: based on this truncation risk strategy, there are two sources of
risk to compare at each step:

1. The (simplified) risk from "all future operations in the sequence".

2. The "direct" risk from the current operation.

The problem is that steps 1 and 2 can't be compared, because in general
they operate on different types.  That's why we need the `UpdateRisk`
utility: to take a risk object, and "back-propagate" it using the
current operation.

Normally, this "back propagation" is very simple: it just updates the
input type.  But sometimes it can be more complex.  For example, if we
have a risk `ValueIsNotInteger<float>`, but it's preceded by a
`StaticCast<int, float>`, then it should become `NoTruncationRisk<int>`,
because we know the `int` input will always produce integral `float`
values.

Helps #349.