Support policy argument for `Quantity` `in`, `as`
This policy will indicate, _at compile time_, which conversion risks we
should guard against.  To implement this, we add a user-facing "risk
set" idea, with constants `OVERFLOW_RISK`, `TRUNCATION_RISK`, and
`ALL_RISKS` (which is `OVERFLOW_RISK | TRUNCATION_RISK`).  Users can
create a _policy_ by passing any of these to either `check()` or
`ignore()`.

So, for example, `inches(36).as(feet, ignore(TRUNCATION_RISK))` will now
produce `feet(3)`.

This gives us a more intent based interface for circumventing safety
checks, as compared to `coerce_as` and `coerce_in`.  We plan to provide
replacements for the other instances of the "coerce" word, which can be
found in `:quantity_point` and `:constant`.  Then, if real-world
experience with these new APIs is encouraging, we plan to formally
deprecate all "coerce" word APIs in 0.6.0, and delete them in 0.7.0.

Helps #387.