Implement `TruncationRiskFor<OpSequence<Ops...>>` (#469)
The risk for a sequence of operations gets propagated backwards from the
last one.  At any given point, there are two competing risks: (a) the
"updated" (i.e., backpropagated) risk from all future operations, and
(b) the "direct" risk from the current operation.  We add an ordering of
risks (via the `TruncationRiskClass` mixin) and a `BiggestRiskImpl`
helper that uses that ordering.

This finishes the `:truncation_risk` library.  Helps #349.