Use trait interface for "mag kind for" (#475)
The Green Hills Compiler complains that `MagKind::DEFAULT` is
unreachable, ostensibly because it is hostile to the idea of generic
programming that supports multiple C++ versions at the same time.  (That
is: I think it _wants_ us to just use `if constexpr` when compiled in
C++17 mode, but we can't because we want to support C++14.)

That's fine.  We can just as easily turn this into a trait.  Defining a
type using nested `std::conditional` constructs lets us avoid misleading
the compiler into thinking that there is any kind of control flow going
on here.

Helps #467.