Support conversion risk policy with `Constant`
`Constant` is a really interesting case for conversion risks, because it
has a perfect conversion policy: we always know at compile time whether
the _actual value_ will be lossy in any way.  Unfortunately, this seems
to make it hard to explicitly opt out of _individual_ risks.

For example, consider getting the speed of light in km/s.  Obviously,
we'll need to provide `ignore(TRUNCATION_RISK)` as a second parameter,
and we would expect this to produce a value of `299'792`.  However, this
won't work, because the overflow risk (which we retained) will be
checking _generic_ values (measured in `c`!), and forbidding any
operation where a value of `2'147` would overflow.  Indeed, it turns out
that `2'147 * c` _would_ overflow `int` in units of km/s!

The solution is amusing: use the _runtime_ conversion checkers, at
_compile_ time!  After all, they are `constexpr` compatible, and here we
know our value at compile time.  This seems like it works really well so
far.  We are able to explicitly specify individual kinds of lossiness
that we're OK with.

Fixes #387.