Split out base power computation into templates (#479)
Those templates are primarily on the sign of `N`, and on whether or not
the type is an integer.

The new implementation of `root` handles 3 common and universally
applicable trivial cases.  Then, it delegates to a so-called "nontrivial
root implementation" type to do the rest.  This implementation splits on
whether the type in question is an integer type (using the simplest yet
most expansive detection that I know), or not.  (We used to split on
`std::is_floating_point`, but that gave us the wrong default behavior
for non-arithmetic types.)

For integers, the "nontrivial root implementation" is simple: we just
return an error condition.  (I added a comment explaining why we don't
have to handle integers whose roots are exact integers.)

For other types, we copy the remaining implementation wholesale: handle
the remaining special cases (negative numbers and numbers in (0, 1)),
and then dive in to the full algorithm.

As for the `base_power_value` implementation itself (which is,
incidentally, the only caller of `root()` in the codebase), we use a new
`enum class` to split based on the sign in a readable way.  Amusingly,
this saves us one `static_cast<std::uintmax_t>` in one branch's
implementation, because now we do it upstream.

Fixes #467.  This was tested on the Aurora-internal repo: we got a fully
clean build when this was combined with #475.

---------

Co-authored-by: Michael Hordijk <hordijk@aurora.tech>