Replace external `UnitAvoidance` with new trait (#483)
We want to stop telling end users to specialize
`::au::detail::UnitAvoidance`, because it's in the `detail` namespace,
which they should never name.  Instead, we provide
`::au::UnitOrderTiebreaker`, which has _almost_ the same effect, but is
in the `::au` namespace and so we can support it.

The difference alluded to above is that, as the name suggests, this is a
_tiebreaker_.  Specializing (the old) `UnitAvoidance`, on the other
hand, affected the _coarse_ ordering of units.  So, we replaced _that_
role with `CoarseUnitOrdering`.

`UnitOrderTiebreaker` will default to `UnitAvoidance` for now, so as to
provide a smooth upgrade path.  In 0.6.0, `UnitAvoidance` will stop
working, and we'll add a deprecation warning to tell people what to do
instead.

Fixes #480.