Add support for {fmt} (#491)
The structure of this support was much more challenging than I had
anticipated.  I _wanted_ to get out-of-the-box support for anyone who
included `"fmt/format.h"`, by forward declaring `fmt::formatter`, and
then specializing it.  This doesn't work because `fmt::formatter` is
actually `fmt::v11::formatter` (currently), where `v11` is an inline
namespace, and `fmt::v10::formatter`, `fmt::v9::formatter`, and so on
are also out in the wild.

The next best thing is to write a class that has _all of the logic_ of
what would be `fmt::formatter<Quantity<U, R>>`, without actually taking
a physical dependency on the fmt library.  This actually does work, and
this is what we do.  End users will need to make their own file which
depends on Au, depends on fmt, and specializes `fmt::formatter` by just
inheriting from `au::QuantityFormatter`.

Now for the actual syntax.  I ended up going with two optional parts:
one for the numeric value, and one for the unit label.  The unit label
gets a `U` prefix, and we use a `;` to separate them if they're both
there, but otherwise the syntax is simply "whatever syntax you would use
for the underlying types".  This gives a lot of flexibility without a
lot of complexity!

I also added documentation explaining how to use the formatting support.

Helps #149.  Really, this does _almost all_ of the work, but we can't
actually test the `std::format` support with our current toolchains (as
we're blocked on #485, #486, and #487).  So the plan is to land the
support for `{fmt}`, and then play around on godbolt to validate our
support for `std::format`.