Make `parse_ctx` constructor call more precise
Here is the first attempt to test `std::format` support:
<https://godbolt.org/z/fYbarPdvG>.

The error complains about our `static_cast<int>` needing to _then_
become a `size_t` in the `std::format` implementation.  So I decided to
look into the constructors in more detail.  It turns out that for
`{fmt}`, the constructor takes a `fmt::string_view` and optional `int`
argument count, while the `std::format` takes a `std::string_view` and a
`size_t` argument count.

So this makes it seem like the constructor was spelled wrongly, and we
were actually providing an argument count, not a second argument to the
`xyz::string_view`.  Oops!

The fix is to construct it with a _nested_ braced list whose second
argument is a `size_t`.  For good measure, we switch to `()` instead of
`{}` to avoid any initializer list confusion.  This should pretty
robustly get us to constructing "whatever kind of string view the
specific library needs".

Helps #149.  After this, we will try again with the godbolt link ---
unfortunately, we can't easily test the `std::format` directly just yet.