Revert support for {fmt}
This implementation is infeasible, because it relies on internal
implementation details of the `{fmt}` library.  We really don't want
this to end up in a public facing release.  See this incredibly
voluminous [comment] for more details.

This PR is a revert of #491, except with the GitHub actions changes
retained, because they are fine, and they might make future changes
easier.

Helps #149, for some definition of "helps".

[comment]: https://github.com/aurora-opensource/au/issues/149#issuecomment-3144905376