Rename `check()` to `guard_against()`
Bad news: `check(TRUNCATION_RISK)` won't work if the user's namespace
has a `check` namespace.  Aurora does; that's how I found out.

Not a huge deal: we can rename `check()` to `guard_against()`.  This
isn't quite as concise, but it reads very nicely, and is extremely
unlikely to be a namespace anywhere out in the wild (which I think is
also true of `ignore`.

Follow-up to #387.