Update docs to reflect conversion risk policies
We add three new documents: two discussion docs on truncation risk and
conversion risks generally, and one reference doc for the conversion
risk policy objects.

Then we go through and update all of the affected docs.  The main types
of changes are:

- Update new APIs (constructors, conversion functions) that can take a
  policy argument.
- Mark the "coerce" APIs as "to be deprecated", and tell people to use
  the policy arguments instead.
- If we find outdated language around permitted/forbidden conversions,
  or around "explicit Rep" APIs, clean it up

This should get the Au docs into a good state with respect to our modern
approach to conversion safety.

Helps #470.