Update docs for common point unit (#501)
The common point unit is simply the common _unit_ of both the input
units, and of the implicitly created "origin displacement" units as
appropriate.  Its origin is the lowest origin of all input units.

This PR fleshes out the description in the discussion page, and provides
a full definition in the reference page.  We also include a worked
example of how the GCD algorithm lets you compute what the common point
unit is.

Finally, I added some links for some usages I saw of "quantity
equivalent" and "point equivalent", because obviously we can't expect
users to just know what these mean.

Helps #470.

---------

Co-authored-by: Michael Hordijk <hordijk@aurora.tech>