Support `{fmt}` and `std::format`
We tried this earlier (#491), but then reverted it (#495).  That was
just a few short weeks ago: what changed?

We learned that our old format string syntax wasn't just _hard_ to
implement; it was actually [impossible].  This led naturally to a [new
idea] that may not be as flexible, but actually still supports the most
important use cases.  Happily, this new idea is very easy to implement
without using any internal implementation details of the formatting
libraries.  It even gives us a little (_very_ little) room for evolving
more options for formatting the unit label, such as setting the fill
character, although we don't currently expect to ever need or want them.

I was also able to verify support for `std::format` by generating a
single file, and uploading it to [godbolt].  This shows that on a
compiler version and setup that includes `std::format`, Au supports it
out of the box with no configuration.

Overall, I feel confident enough in this solution to land it for the
0.5.0 release.  Fixes #149.

[impossible]: https://github.com/fmtlib/fmt/issues/4508#issuecomment-3172708735
[new idea]: https://github.com/fmtlib/fmt/issues/4508#issuecomment-3172783356
[godbolt]: https://godbolt.org/z/6jhW1jboW