Add `std::format`-supporting single-file headers
These will now be automatically generated and included in the
installation guide.  This effectively makes each "group" a 2x2 matrix,
instead of just having 2 values (yes/no for `<iostream>`).  So I changed
the display to a matrix format to be more visual.

Yes, this does double the number of single-file versions that we
publish.  The increase in flexibility far outweighs the miniscule added
cost to generate and host these, IMO.  That said, we probably cannot
support _too_ many more doublings. :sweat_smile:

Tested with `au-docs-serve`.  Note that the relative paths change from
`./` to `../`.  Formerly, we were using markdown syntax, and "relative"
meant "relative to `install.md`".  Now, in HTML, it means "relative to
`install/` on the output website": `install` becomes a _directory_, so
we need `../` to go up a level and find the files.

Helps #470.