Update documentation of library alternatives
First, we do a version bump on all libraries, so that we're comparing to
the latest versions.

Next, our delivery information is outdated.  The single file delivery is
no longer our greatest strength.  Instead, we should emphasize our full
support for bazel and CMake, and community support for conan and vcpkg.

Now to the features.

There is a brand new one: negative units.  As far as I know, nobody
supports these except for us.  I know for a fact that mp-units does not,
and I checked the others' documentation as best I could.

On I/O, now that we support fmtlib and `std::format`, we upgrade
ourselves to "good".  I think mp-units is still the best, because they
have a lot of customizability on the unit label (e.g., toggling between
"pure ASCII" solutions and using special characters).  Their format
string support has a lot more customization too, although I don't think
most of the differences matter much.  I'm downgrading boost units to
"fair", because they don't support format strings.  This gives us a nice
consistent approach: you can't be "good" if you don't support format
strings, which feels right for 2025 C++.

On point types, we now have maximally efficient common units, which is a
great feature that nobody else has.  I'm still rating mp-units as the
best, because of their more mature support for custom origins.
Additionally, they switched their approach since the last time we rated
them to use `point<U>` and `delta<U>` modifiers that apply directly to a
quantity spec `U`, which is a nice and elegant approach, so we reflect
that in our assessment.

Finally, for conversion safety, I now give us credit for having
separable checks for overflow and truncation.  We were already easily
the best, and now we're even better.

Helps #470.