Clarify how-to docs for new units
Some users have found some aspects of these docs confusing.  This PR
fixes them up to make the guide easier to use.  Here are the problems
and how we handled them.

- **Problem:** It was unclear which namespace should contain the
  definitions of new units.
    - **Solution:** Append `, in your project's namespace` to the
      comments in the guide.

- **Problem:** It was unclear which namespace the `symbols` namespace
  should go in.
    - **Solution:** The above change should automatically clarify this
      as well.

- **Problem:** The `au::` namespace was omitted from Au library
  constructs, making it unclear whether we had elided a `using` clause,
  or whether we wanted users to define the units in the `au` namespace.
    - **Solution:** The desired namespace should no longer be unclear.
      We had been writing the docs by omitting the `au` namespace to
      reduce visual clutter, but this is less appropriate for a how-to
      guide.  We have added the namespace.

Of course, this last point applies with equal force to every other
how-to guide.  It's OK to eliminate the `au::` namespace qualifier in
the _text_, but we should keep the _code samples_ as usable as possible.
Therefore, we apply this approach more consistently.

Fixes #504.