Support `{fmt}` and `std::format` (#505)
We tried this earlier (#491), but then reverted it (#495).  That was
just a few short weeks ago: what changed?

We learned that our old format string syntax wasn't just _hard_ to
implement; it was actually [impossible].  This led naturally to a [new
idea] that may not be as flexible, but actually still supports the most
important use cases.  Happily, this new idea is very easy to implement
without using any internal implementation details of the formatting
libraries.  It even gives us a little (_very_ little) room for evolving
more options for formatting the unit label, such as setting the fill
character, although we don't currently expect to ever need or want them.

I was also able to verify support for `std::format` by generating a
single file, and uploading it to [godbolt].  This shows that on a
compiler version and setup that includes `std::format`, Au supports it
via the `"au/std_format.hh"` header.  I also updated the single-file
script to have `--no-std-format` (which is the default) and
`--std-format` options.  A future PR will add versions with/without
`std::format` support to our automatically generated single-file
options.

On the CMake side, we need to skip linting for `std_format.hh`.  It's
too hard and too complicated to set up reliable rules for when we should
and shouldn't make it available.  Therefore, we _always_ make it
available, but we trust that users will only include it if they know
they can support it.  That's the point of the opt-in header strategy.
Note, though, that this required us to bump the minimum CMake version
from 3.24 to 3.27.

Fixes #149.

[impossible]: fmtlib/fmt#4508 (comment)
[new idea]: fmtlib/fmt#4508 (comment)
[godbolt]: https://godbolt.org/z/f6Pvz9Mzb

---------

Co-authored-by: Michael Hordijk <hordijk@aurora.tech>