Circumvent more Green Hills compiler warnings (#515)
For `get_value_result()`, we split the implementation between two new
types: one for "trying to put a non-integer in an integral type", and
one for "everything else".  We still assign the result to a `constexpr`
variable to force it to be evaluated at compile time.

For the other two, we use ternary operators to bring things into a
single line of code, splitting out into separate functions when that
makes sense.

Follow-on to #467.